experimental/lapis/pyproject.toml
[project]
name = "lapis"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = ["numpy", "PyMCubes", "taichi"]